give WranglerToml structs their own file
